Removing JSON gem compatibility layer
This causes issues because JSON and ffi-yajl get into fights over who
monkeypatched the classes last and this causes random issues with
whitespace disagreement between the JSON gem and the underlying libyajl
C library.  For deterministic output its better not to have fighting
monkeypatches, it generates considerable chaos.

If anyone wants to extract the json_gem.rb file into a gem of its own
and support these monkeypatches going forwards that would be fine
(there is, in fact, literally nothing that could be done to stop you),
but we do not want to support these.