Remove ffi gem as a hard dependency
Copies the algorithm out of FFI.map_library_name into the
FFI_Yajl#map_library_name function so that when we're loading up the
C extension we don't need to also slurp in the ffi gem just to run
that.  The ffi gem is moved to a hard dependency of only the jruby
version of the gem, and its a development dependency for the other
ruby VMs (so rake spec will work against the ffi library), but its
an optional library to install to make the ffi library work on VMs
that support C extensions (MRI, RBX).