Use versioned .so file.
Using non-versioned .so file is wrong. This file is used intended to be
used just by linker. This file is typically just symlink to the
versioned .so file. Most distributions ships this file in -devel
subpackage and it is not installed on user system, nor it is required
for runtime.

Also, ffi_yajl depends on yajl version 2 and no other, so requiring the
versioned .so file is correct also from this POV.

Signed-off-by: Vít Ondruch <vondruch@redhat.com>